minor edit 🧞 👨‍💻 🦹 🔥

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -13,7 +13,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t index;
+	unsigned int index;
 
 	if (array != NULL && size > 0 && action != NULL)
 		for (index = 0; index < size; index++)
